unskip integration tests for INDY-2262; allow processing checkpoints for future views
