+++ plenum/test/consensus/order_service/test_ordering_service_on_view_change.py
@@ -6,7 +6,6 @@ from plenum.common.util import updateNamedTuple
 from plenum.server.consensus.consensus_shared_data import preprepare_to_batch_id
 from plenum.server.consensus.batch_id import BatchID
 from plenum.server.consensus.ordering_service import OrderingService
-from plenum.server.consensus.ordering_service_msg_validator import OrderingServiceMsgValidator
 from plenum.server.future_primaries_batch_handler import FuturePrimariesBatchHandler
 from plenum.server.replica_helper import generateName
 from plenum.test.consensus.helper import copy_shared_data, create_batches, \
@@ -37,7 +36,6 @@ def orderer(_orderer, is_primary, ):
                                                                          primaries=_orderer._data.primaries))
     write_manager.register_batch_handler(future_primaries_handler)
 
-    _orderer._validator = OrderingServiceMsgValidator(_orderer._data)
     _orderer.name = 'Alpha:0'
     _orderer._data.primary_name = 'some_node:0' if not is_primary else orderer.name
