+++ plenum/test/view_change_service/test_view_change_while_adding_new_node_2_slow_commit.py
@@ -1,9 +1,6 @@
-import pytest
-
 from plenum.test.view_change_service.helper import check_view_change_adding_new_node
 
 
-@pytest.mark.skip('INDY-2262')
 def test_view_change_while_adding_new_node_2_slow_commit(looper, tdir, tconf, allPluginsPath,
                                                          txnPoolNodeSet,
                                                          sdk_pool_handle,
@@ -17,4 +14,4 @@ def test_view_change_while_adding_new_node_2_slow_commit(looper, tdir, tconf, al
                                       slow_nodes=[txnPoolNodeSet[1], txnPoolNodeSet[2]],
                                       delay_pre_prepare=False,
                                       delay_commit=True,
-                                      expected_viewno=5)
+                                      expected_viewno=4)
